Revised install_tbb.sh such that TBB's RT-lib is updated according to the TBB version.

+++ contrib/libs/tbb/install_tbb.sh
@@ -1,27 +1,29 @@
 #!/bin/bash
 
+BASENAME=$(command -v basename)
 DIRNAME=$(command -v dirname)
 SORT=$(command -v sort)
 TAIL=$(command -v tail)
-ENV=$(command -v env)
+SED=$(command -v sed)
+LS=$(command -v ls)
 RM=$(command -v rm)
 CP=$(command -v cp)
 
 TBBPREFIX="/opt/intel /usr/local ${HOME}"
 HERE=$(cd $(${DIRNAME} $0); pwd -P)
 
-if [ "" != "${DIRNAME}" ] && \
-   [ "" != "${SORT}" ] && [ "" != "${TAIL}" ] && \
-   [ "" != "${RM}" ] && [ "" != "${CP}" ];
+if [ "" != "${SORT}" ] && [ "" != "${TAIL}" ] && [ "" != "${SED}" ] && \
+   [ "" != "${LS}" ] && [ "" != "${RM}" ] && [ "" != "${CP}" ] && \
+   [ "" != "${BASENAME}" ] && [ "" != "${DIRNAME}" ];
 then
   for DIR in ${TBBPREFIX}; do
     if [ "" != "${TBBROOT}" ]; then break; fi
-    TBBHEAD=$(ls -1 ${DIR}/compilers_and_libraries_*/linux/tbb/include/tbb/tbb.h 2>/dev/null | ${SORT} -V | ${TAIL} -n1)
+    TBBHEAD=$(${LS} -1 ${DIR}/compilers_and_libraries_*/linux/tbb/include/tbb/tbb.h 2>/dev/null | ${SORT} -V | ${TAIL} -n1)
     TBBROOT=$(${DIRNAME} $(${DIRNAME} $(${DIRNAME} ${TBBHEAD} 2>/dev/null) 2>/dev/null) 2>/dev/null)
   done
   for DIR in ${TBBPREFIX}; do
     if [ "" != "${TBBROOT}" ]; then break; fi
-    TBBHEAD=$(ls -1 ${DIR}/tbb*/include/tbb/tbb.h 2>/dev/null | ${SORT} -V | ${TAIL} -n1)
+    TBBHEAD=$(${LS} -1 ${DIR}/tbb*/include/tbb/tbb.h 2>/dev/null | ${SORT} -V | ${TAIL} -n1)
     TBBROOT=$(${DIRNAME} $(${DIRNAME} $(${DIRNAME} ${TBBHEAD} 2>/dev/null) 2>/dev/null) 2>/dev/null)
   done
   echo "Intel TBB found at ${TBBROOT}."
@@ -43,8 +45,15 @@ then
         echo -n " delete old directory lib, and"
         ${RM} -r ${HERE}/lib
       fi
-      echo " depp-copy ${TBBROOT}/lib"
+      echo " deep-copy ${TBBROOT}/lib"
       ${CP} -Lr ${TBBROOT}/lib ${HERE}
+      TBBRTNEW=$(${LS} -1 ${HERE}/lib/intel64 2>/dev/null | ${SORT} -V | ${TAIL} -n1)
+      TBBRTCUR=$(${BASENAME} $(${SED} -n "s/ *-L\(..*\)/\1/p" ${HERE}/ya.make.template))
+      if [ "${TBBRTNEW}" = "${TBBRTCUR}" ]; then
+        ${CP} ${HERE}/ya.make.template ${HERE}/ya.make
+      else
+        ${SED} -e "s/\(.*\)-L\(..*\)${TBBRTCUR}/\1-L\2${TBBRTNEW}/" ${HERE}/ya.make.template > ${HERE}/ya.make
+      fi
       echo "Successfully completed."
     else
       echo "No action performed."
